[SAY-THINK] Include basic speech_manager.hpp class if no speech support
Avoids issues where the included classes add platform-specific implementations which could cause compilation errors. With the basic speech_manager included, nothing will happen if the platform's render.cpp does not try to render speech (intended). IMPORTANT NOTE: It is possible to implement the renderer for a platform and not add it to the ifdefs and it will still work, but could cause runtime crashes if platform-specific functions are called in shared files. This shouldn't happen though and hopefully we'll have a unified rendering API before we need to worry about that.